// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/create-content-report-http-request.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateContentReportHttpRequest {
    reason?: string;
    targetId?: string;
    targetType?: string;
}