// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/recipe-comment-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface RecipeCommentHttpResponse {
    authorFirstName?: string | null;
    authorId?: string;
    authorUsername?: string | null;
    createdAtUtc?: string;
    id?: string;
    isOwnedByCurrentUser?: boolean;
    modifiedAtUtc?: string | null;
    recipeId?: string;
    text?: string;
}